requirements.txt
streamlit~=1.45.1
python-dotenv~=1.1.0
requests~=2.32.3
chromadb~=1.0.10
tabulate~=0.9.0